pyproject.toml
[project]
name = "ptsd-homework-agent"
version = "0.1.0"
description = "PTSD 심리치료 대화 지원 에이전트 - 대화 기술 및 정서 표현 도움"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # 핵심 프레임워크
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    # LangChain AI 모델 통합 (단순화)
    "langchain-openai>=0.2.0",
    # Firebase 연동
    "firebase-admin>=6.5.0",
    # 환경변수 및 설정
    "python-dotenv>=1.0.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    # 유틸리티
    "aiofiles>=24.1.0",
    "typing-extensions>=4.12.0",
    # 로깅
    "loguru>=0.7.0",
    "streamlit>=1.46.1",
    # 데이터 시각화 (대시보드용)
    "pandas>=2.0.0",
    "plotly>=5.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.8.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
